refactor: MVC 기반으로 폴더구조 변경
