+++ src/main/java/step5/model/MoveStrategy.java
@@ -1,4 +1,4 @@
-package step5;
+package step5.model;
 
 public interface MoveStrategy {
     Position move(Position position);
