+++ src/main/java/step5/model/RandomMoveStrategy.java
@@ -1,4 +1,4 @@
-package step5;
+package step5.model;
 
 import java.util.Random;
