+++ src/main/java/step5/view/GameOutput.java
@@ -1,6 +1,7 @@
-package step5;
+package step5.view;
 
 import java.util.List;
+import step5.model.CarRacingGame;
 
 public class GameOutput {
