+++ src/test/java/step5/CarRacingGameTest.java
@@ -5,6 +5,7 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+import step5.model.CarRacingGame;
 
 public class CarRacingGameTest {
