+++ src/test/java/step5/PositionTest.java
@@ -4,6 +4,8 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import org.junit.jupiter.api.Test;
+import step5.model.InvalidPositionException;
+import step5.model.Position;
 
 public class PositionTest {
